feat: Fall-back to "search for user by display name"
In case the build email is missing due to too old version of Drone server is being used.

AFAIK, `p.Build.Email` is not set in Drone v0.8.6. I know it is 2 minor versions old compared to the latest release, but this feature breaks nothing while making more users happy :)